Refactor: Image 유틸 클래스를 interface로 변경
