+++ src/main/java/com/nhnacademy/marketgg/server/service/impl/StorageImageService.java
@@ -1,0 +1,52 @@
+package com.nhnacademy.marketgg.server.service.impl;
+
+import com.nhnacademy.marketgg.server.entity.Asset;
+import com.nhnacademy.marketgg.server.entity.Image;
+import com.nhnacademy.marketgg.server.service.ImageService;
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+import java.util.UUID;
+import org.springframework.stereotype.Component;
+import org.springframework.util.CollectionUtils;
+import org.springframework.web.multipart.MultipartFile;
+
+@Component
+public class StorageImageService implements ImageService {
+
+    private static final String DIR = System.getProperty("user.home");
+
+    @Override
+    public List<Image> parseImages(List<MultipartFile> multipartFiles, Asset asset) throws IOException {
+        List<Image> images = new ArrayList<>();
+
+        if (!CollectionUtils.isEmpty(multipartFiles)) {
+            String dir = String.valueOf(Files.createDirectories(returnDir()));
+            Integer sequence = 1;
+
+            for (MultipartFile multipartFile : multipartFiles) {
+
+                sequence++;
+            }
+        }
+        return images;
+    }
+
+    private Path returnDir() {
+        String format = LocalDate.now().format(DateTimeFormatter.ISO_LOCAL_DATE);
+
+        return Paths.get(DIR, format);
+    }
+
+    private String uuidFilename(String filename) {
+
+        return UUID.randomUUID() + "_" + filename;
+    }
+}
